Remove retired StoreHub live integration
StoreHub POS is fully retired (final outlet cut over to POS-native on
2026-06-18; account cancelled 2026-06-27). Remove the dead live-integration
code while preserving the read-only storehub_sales / storehub_products
archive and all historical reporting that reads it.

Deleted (live StoreHub API — now non-functional):
- lib/storehub.ts — the StoreHub BackOffice API client
- lib/storehub-archive.ts + cron/storehub-sync — the archive freshness sync
  (already de-scheduled in vercel.json; archive is frozen now)
- inventory/storehub/sync-sales + sync-products — live catalogue/sales sync
- finance/ingestors/storehub-eod.ts + finance/ingest/storehub-eod route —
  the live StoreHub AR/EOD ingestor and its manual trigger
- sales/debug, sales/recommendations, sales/targets/recompute — sales
  routes built entirely on the live StoreHub API
- STOREHUB_ACCOUNT_ID / STOREHUB_API_KEY from .env.example

Kept (historical archive — read-only, the point of the retention):
- storehub_sales / storehub_sale_items / storehub_products tables
- SalesTransaction archive (still read by inventory/COGS/cashflow/projection)
- storehub-helpers classification, staff dashboard bridge, reports, compare,
  command center, and finance-eod-backfill

Edited:
- unified-sales.ts: dropped the live "today" pull; today now reads the same
  frozen archive, cutover de-dup gate preserved. Relocated the
  StoreHubTransaction type into storehub-helpers (its old home was deleted).
- pos-native-eod.ts: moved IngestEodResult in-file; ingestEodForDate is now
  fully native (dropped the pre-cutover StoreHub routing branch).
- sales dashboard page: removed the AI Sales Coach panel and the targets
  Recompute button (both powered by the deleted routes).
