+++ .env.example
@@ -10,14 +10,6 @@ AUTH_COOKIE_DOMAIN=".celsiuscoffee.com"
 NEXT_PUBLIC_SUPABASE_URL="https://xxx.supabase.co"
 NEXT_PUBLIC_SUPABASE_ANON_KEY="eyJ..."
 
-# ── StoreHub (LEGACY — POS migrated to native) ───
-# Only used to read historical sales/products during the StoreHub → native
-# migration. Safe to remove (here and from Vercel: backoffice + loyalty) once
-# StoreHub is retired — target 2026-06-27. Loyalty used STOREHUB_USERNAME for
-# the same account id; collapse to STOREHUB_ACCOUNT_ID when removing.
-STOREHUB_ACCOUNT_ID=""
-STOREHUB_API_KEY=""
-
 # ── Indeed Sponsored Jobs (recruitment ad spend) ─
 # Register app at https://secure.indeed.com/account/apikeys (OAuth 2.0,
 # client_credentials grant). Confidential client only — never expose to
