+++ apps/backoffice/src/app/(admin)/sales/dashboard/page.tsx
@@ -10,10 +10,7 @@ import {
   UtensilsCrossed,
   ShoppingBag,
   Truck,
-  Sparkles,
   AlertTriangle,
-  Lightbulb,
-  Zap,
 } from "lucide-react";
 import { cn } from "@/lib/utils";
 import { AccumulativeChart } from "./AccumulativeChart";
@@ -100,14 +97,6 @@ type DashboardData = {
   warnings?: string[];
 };
 
-type Recommendation = {
-  type: "opportunity" | "warning" | "insight" | "action";
-  title: string;
-  description: string;
-  impact: "high" | "medium" | "low";
-  category: string;
-};
-
 // ─── Helpers ────────────────────────────────────────────────────────────
 
 function formatRM(n: number): string {
@@ -190,51 +179,9 @@ export default function SalesDashboard() {
   // Active metric tab for the grid
   const [activeMetric, setActiveMetric] = useState<"revenue" | "orders" | "aov">("revenue");
 
-  // AI recommendations
-  const [recommendations, setRecommendations] = useState<Recommendation[]>([]);
-  const [aiLoading, setAiLoading] = useState(false);
-  const [aiError, setAiError] = useState<string | null>(null);
-
-  // AI targets recompute
-  const [recomputingTargets, setRecomputingTargets] = useState(false);
-  const [targetsRecomputeError, setTargetsRecomputeError] = useState<string | null>(null);
+  // Targets panel: toggle for the stored AI reasoning (data comes from the
+  // dashboard route; StoreHub-fed recompute was retired with StoreHub).
   const [showReasoning, setShowReasoning] = useState(false);
-
-  const recomputeTargets = useCallback(async () => {
-    setRecomputingTargets(true);
-    setTargetsRecomputeError(null);
-    try {
-      const res = await fetch("/api/sales/targets/recompute", {
-        method: "POST",
-        credentials: "include",
-      });
-      if (!res.ok) {
-        const body = await res.json().catch(() => ({}));
-        throw new Error(body.error || `HTTP ${res.status}`);
-      }
-      // Trigger dashboard reload to pick up new targets
-      window.location.reload();
-    } catch (err) {
-      setTargetsRecomputeError(err instanceof Error ? err.message : "Failed to recompute");
-      setRecomputingTargets(false);
-    }
-  }, []);
-
-  const loadRecommendations = useCallback(async (outlet: string) => {
-    setAiLoading(true);
-    setAiError(null);
-    try {
-      let url = "/api/sales/recommendations";
-      if (outlet !== "all") url += `?outletId=${outlet}`;
-      const res = await fetch(url, { credentials: "include" });
-      if (!res.ok) throw new Error("Failed to load");
-      const body = await res.json();
-      setRecommendations(body.recommendations || []);
-    } catch {
-      setAiError("Could not load AI recommendations");
-    }
-    setAiLoading(false);
-  }, []);
 
   const loadData = useCallback(
     async (p: Period, outlet: string, cFrom?: string, cTo?: string) => {
@@ -267,11 +214,6 @@ export default function SalesDashboard() {
       loadData(period, outletId);
     }
   }, [period, outletId, customFrom, customTo, loadData]);
-
-  // Auto-load AI insights when outlet changes (debounced mount)
-  useEffect(() => {
-    loadRecommendations(outletId);
-  }, [outletId, loadRecommendations]);
 
   // ─── Render ─────────────────────────────────────────────────────────
 
@@ -495,7 +437,7 @@ export default function SalesDashboard() {
                       </div>
                       <p className="text-[11px] text-gray-500 mt-0.5">
                         {isDefault
-                          ? "No AI-set targets yet. Click Recompute to have AI set progressive targets from your trailing 28-day performance."
+                          ? "No AI-set targets yet — showing defaults."
                           : "Targets auto-scale from trailing actuals. Only move up — never down."}
                       </p>
                       {tm?.reasoning && showReasoning && (
@@ -503,141 +445,22 @@ export default function SalesDashboard() {
                           {tm.reasoning.split(" | ")[0]}
                         </p>
                       )}
-                      {targetsRecomputeError && (
-                        <p className="text-[11px] text-red-500 mt-1">{targetsRecomputeError}</p>
-                      )}
                     </div>
                   </div>
-                  <div className="flex items-center gap-2 shrink-0">
-                    {tm?.reasoning && (
+                  {tm?.reasoning && (
+                    <div className="flex items-center gap-2 shrink-0">
                       <button
                         onClick={() => setShowReasoning((s) => !s)}
                         className="text-[11px] text-gray-500 hover:text-gray-700 underline underline-offset-2"
                       >
                         {showReasoning ? "Hide reasoning" : "Why"}
                       </button>
-                    )}
-                    <button
-                      onClick={recomputeTargets}
-                      disabled={recomputingTargets}
-                      className={cn(
-                        "flex items-center gap-1.5 rounded-lg px-3 py-1.5 text-xs font-medium transition-colors",
-                        recomputingTargets
-                          ? "bg-gray-100 text-gray-400 cursor-not-allowed"
-                          : "bg-white border border-gray-200 text-gray-700 hover:bg-gray-50",
-                      )}
-                    >
-                      {recomputingTargets ? (
-                        <>
-                          <Loader2 className="h-3.5 w-3.5 animate-spin" />
-                          Recomputing...
-                        </>
-                      ) : (
-                        <>
-                          <Sparkles className="h-3.5 w-3.5" />
-                          {isDefault ? "Set Targets" : "Recompute"}
-                        </>
-                      )}
-                    </button>
-                  </div>
+                    </div>
+                  )}
                 </div>
               </div>
             );
           })()}
-
-          {/* ─── AI Recommendations (moved to top) ─── */}
-          <div className="rounded-xl border border-gray-200 bg-white shadow-sm p-5">
-            <div className="flex items-center justify-between mb-4">
-              <div className="flex items-center gap-2">
-                <Sparkles className="h-5 w-5 text-[#C2452D]" />
-                <h3 className="text-sm font-semibold text-gray-900">AI Sales Coach</h3>
-                <span className="text-[10px] bg-[#C2452D]/10 text-[#C2452D] px-1.5 py-0.5 rounded font-medium">
-                  Beta
-                </span>
-              </div>
-              <button
-                onClick={() => loadRecommendations(outletId)}
-                disabled={aiLoading}
-                className={cn(
-                  "flex items-center gap-1.5 rounded-lg px-3 py-1.5 text-xs font-medium transition-colors",
-                  aiLoading
-                    ? "bg-gray-100 text-gray-400 cursor-not-allowed"
-                    : "bg-white text-gray-600 border border-gray-200 hover:bg-gray-50",
-                )}
-              >
-                {aiLoading ? (
-                  <>
-                    <Loader2 className="h-3.5 w-3.5 animate-spin" />
-                    Analyzing...
-                  </>
-                ) : (
-                  <>
-                    <Sparkles className="h-3.5 w-3.5" />
-                    Refresh
-                  </>
-                )}
-              </button>
-            </div>
-
-            {aiError && (
-              <p className="text-xs text-red-500 mb-3">{aiError}</p>
-            )}
-
-            {aiLoading && recommendations.length === 0 && (
-              <div className="flex items-center justify-center py-8 text-gray-400">
-                <Loader2 className="h-5 w-5 animate-spin mr-2" />
-                <span className="text-xs">Reading last 30 days of sales…</span>
-              </div>
-            )}
-
-            {recommendations.length > 0 && (
-              <div className="grid grid-cols-1 sm:grid-cols-2 gap-3">
-                {recommendations.map((rec, idx) => {
-                  const iconMap: Record<string, React.ReactNode> = {
-                    opportunity: <Sparkles className="h-4 w-4 text-green-500" />,
-                    warning: <AlertTriangle className="h-4 w-4 text-amber-500" />,
-                    insight: <Lightbulb className="h-4 w-4 text-blue-500" />,
-                    action: <Zap className="h-4 w-4 text-purple-500" />,
-                  };
-                  const borderMap: Record<string, string> = {
-                    opportunity: "border-l-green-400",
-                    warning: "border-l-amber-400",
-                    insight: "border-l-blue-400",
-                    action: "border-l-purple-400",
-                  };
-                  const impactBadge: Record<string, string> = {
-                    high: "bg-red-50 text-red-600",
-                    medium: "bg-yellow-50 text-yellow-600",
-                    low: "bg-gray-50 text-gray-500",
-                  };
-                  return (
-                    <div
-                      key={idx}
-                      className={cn(
-                        "rounded-lg border border-gray-100 border-l-4 p-3",
-                        borderMap[rec.type] || "border-l-gray-300",
-                      )}
-                    >
-                      <div className="flex items-start gap-2">
-                        <div className="mt-0.5 shrink-0">
-                          {iconMap[rec.type] || <Lightbulb className="h-4 w-4 text-gray-400" />}
-                        </div>
-                        <div className="flex-1 min-w-0">
-                          <div className="flex items-center gap-2 mb-1">
-                            <p className="text-xs font-semibold text-gray-900 truncate">{rec.title}</p>
-                            <span className={cn("text-[10px] px-1.5 py-0.5 rounded font-medium shrink-0", impactBadge[rec.impact] || impactBadge.low)}>
-                              {rec.impact}
-                            </span>
-                          </div>
-                          <p className="text-[11px] text-gray-500 leading-relaxed">{rec.description}</p>
-                        </div>
-                      </div>
-                    </div>
-                  );
-                })}
-              </div>
-            )}
-          </div>
 
           {/* ─── Metric Toggle ─── */}
           <div className="flex items-center gap-2">
