--- apps/backoffice/src/app/api/sales/recommendations/route.ts
@@ -1,363 +1,0 @@
-import { NextRequest, NextResponse } from "next/server";
-import Anthropic from "@anthropic-ai/sdk";
-import { getSession } from "@/lib/auth";
-import { prisma } from "@/lib/prisma";
-import { getTransactions, getProducts, type StoreHubTransaction } from "@/lib/storehub";
-import { getActiveTargets } from "../_lib/targets";
-
-const client = new Anthropic({ apiKey: process.env.ANTHROPIC_API_KEY });
-
-// ─── Time Rounds (MYT = UTC+8) ──────────────────────────────────────────
-
-const ROUNDS = [
-  { key: "breakfast", label: "Breakfast", startH: 8, endH: 10 },
-  { key: "brunch", label: "Brunch", startH: 10, endH: 12 },
-  { key: "lunch", label: "Lunch", startH: 12, endH: 15 },
-  { key: "midday", label: "Midday", startH: 15, endH: 17 },
-  { key: "evening", label: "Evening", startH: 17, endH: 19 },
-  { key: "dinner", label: "Dinner", startH: 19, endH: 21 },
-  { key: "supper", label: "Supper", startH: 21, endH: 23 },
-] as const;
-
-function isWeekend(dateStr: string): boolean {
-  const d = new Date(dateStr + "T12:00:00+08:00");
-  const day = d.getDay();
-  return day === 0 || day === 6;
-}
-
-/**
- * Blended daily target across a set of dates, using active DB targets
- * (weekday vs weekend depending on each date).
- */
-function getAvgTarget(
-  roundKey: string,
-  dates: string[],
-  activeTargets: Awaited<ReturnType<typeof getActiveTargets>>["targets"],
-): number {
-  if (dates.length === 0) return 0;
-  const t = activeTargets[roundKey as keyof typeof activeTargets];
-  if (!t) return 0;
-  let total = 0;
-  for (const d of dates) {
-    total += isWeekend(d) ? t.weekend.revenue : t.weekday.revenue;
-  }
-  return Math.round(total / dates.length);
-}
-
-function getMYTHour(dateStr: string): number {
-  const d = new Date(dateStr);
-  if (isNaN(d.getTime())) return -1;
-  const isUTC = /Z|[+-]\d{2}:\d{2}$/.test(dateStr);
-  if (isUTC) {
-    const myt = new Date(d.getTime() + 8 * 60 * 60 * 1000);
-    return myt.getUTCHours();
-  }
-  return d.getUTCHours();
-}
-
-function getMYTDateStr(dateStr: string): string {
-  const d = new Date(dateStr);
-  if (isNaN(d.getTime())) return "unknown";
-  const isUTC = /Z|[+-]\d{2}:\d{2}$/.test(dateStr);
-  if (isUTC) {
-    const myt = new Date(d.getTime() + 8 * 60 * 60 * 1000);
-    return myt.toISOString().split("T")[0];
-  }
-  return d.toISOString().split("T")[0];
-}
-
-function getRoundKey(hour: number): string | null {
-  for (const r of ROUNDS) {
-    if (hour >= r.startH && hour < r.endH) return r.key;
-  }
-  return null;
-}
-
-function classifyChannel(channel?: string | null): "dine_in" | "takeaway" | "delivery" {
-  if (!channel) return "dine_in";
-  const lower = channel.toLowerCase().trim();
-  if (lower === "takeaway" || lower === "take-away" || lower === "take away") return "takeaway";
-  if (["delivery", "grab", "grabfood", "foodpanda", "shopee", "shopeefood"].includes(lower)) return "delivery";
-  return "dine_in";
-}
-
-// ─── GET /api/sales/recommendations ────────────────────────────────────
-
-export async function GET(request: NextRequest) {
-  try {
-    const user = await getSession();
-    if (!user) {
-      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
-    }
-
-    const { searchParams } = new URL(request.url);
-    const outletId = searchParams.get("outletId") || null;
-
-    // Load active (AI-set) targets for use in % calculations and the prompt
-    const { targets: activeTargets, meta: targetsMeta } = await getActiveTargets();
-
-    // Get date ranges: last 7 days and last 30 days
-    const now = new Date();
-    const mytNow = new Date(now.getTime() + 8 * 60 * 60 * 1000);
-    const todayMYT = mytNow.toISOString().split("T")[0];
-
-    const d7 = new Date(mytNow);
-    d7.setDate(d7.getDate() - 6);
-    const from7 = d7.toISOString().split("T")[0];
-
-    const d30 = new Date(mytNow);
-    d30.setDate(d30.getDate() - 29);
-    const from30 = d30.toISOString().split("T")[0];
-
-    // Fetch outlets
-    const outletWhere = outletId
-      ? { id: outletId, storehubId: { not: null } }
-      : { storehubId: { not: null }, status: "ACTIVE" as const };
-
-    const outlets = await prisma.outlet.findMany({
-      where: outletWhere,
-      select: { id: true, name: true, storehubId: true },
-    });
-
-    if (outlets.length === 0) {
-      return NextResponse.json({ error: "No outlets with StoreHub configured" }, { status: 404 });
-    }
-
-    // Fetch last 30 days of transactions
-    const allTxns: StoreHubTransaction[] = [];
-    for (const outlet of outlets) {
-      if (!outlet.storehubId) continue;
-      try {
-        const from = new Date(from30 + "T00:00:00+08:00");
-        const to = new Date(todayMYT + "T23:59:59+08:00");
-        const txns = await getTransactions(outlet.storehubId, from, to);
-        allTxns.push(...txns);
-      } catch (err) {
-        console.error(`[sales/recommendations] Failed to fetch for outlet ${outlet.name}:`, err);
-      }
-    }
-
-    if (allTxns.length === 0) {
-      return NextResponse.json({ recommendations: [], summary: "No sales data available." });
-    }
-
-    // StoreHub transaction items carry productId, NOT name.
-    // Fetch the product catalog once to resolve names (same pattern as sync-sales).
-    const shProducts = await getProducts().catch(() => []);
-    const productNameById = new Map(shProducts.map((p) => [p.id, p.name]));
-
-    // ─── Aggregate data for AI analysis ─────────────────────────────
-
-    // 1. Round performance (last 7 days vs last 30 days)
-    type RoundStats = { revenue: number; orders: number; days: Set<string> };
-    const roundStats7: Record<string, RoundStats> = {};
-    const roundStats30: Record<string, RoundStats> = {};
-    for (const r of ROUNDS) {
-      roundStats7[r.key] = { revenue: 0, orders: 0, days: new Set() };
-      roundStats30[r.key] = { revenue: 0, orders: 0, days: new Set() };
-    }
-
-    // 2. Product performance (global + per-round)
-    type ProductStats = { name: string; revenue: number; quantity: number; orders: number };
-    const productStats: Record<string, ProductStats> = {};
-    const productStatsByRound: Record<string, Record<string, ProductStats>> = {};
-    for (const r of ROUNDS) productStatsByRound[r.key] = {};
-
-    // 3. Channel breakdown
-    const channelStats = { dine_in: { revenue: 0, orders: 0 }, takeaway: { revenue: 0, orders: 0 }, delivery: { revenue: 0, orders: 0 } };
-
-    // 4. Daily totals for trend
-    const dailyTotals: Record<string, { revenue: number; orders: number }> = {};
-
-    for (const txn of allTxns) {
-      const ts = txn.transactionTime || txn.completedAt || txn.createdAt;
-      if (!ts) continue;
-
-      const hour = getMYTHour(ts);
-      const dateStr = getMYTDateStr(ts);
-      const round = getRoundKey(hour);
-      const channel = classifyChannel(txn.channel);
-
-      // Daily
-      if (!dailyTotals[dateStr]) dailyTotals[dateStr] = { revenue: 0, orders: 0 };
-      dailyTotals[dateStr].revenue += txn.total;
-      dailyTotals[dateStr].orders += 1;
-
-      // Channel
-      channelStats[channel].revenue += txn.total;
-      channelStats[channel].orders += 1;
-
-      // Round
-      if (round) {
-        roundStats30[round].revenue += txn.total;
-        roundStats30[round].orders += 1;
-        roundStats30[round].days.add(dateStr);
-        if (dateStr >= from7) {
-          roundStats7[round].revenue += txn.total;
-          roundStats7[round].orders += 1;
-          roundStats7[round].days.add(dateStr);
-        }
-      }
-
-      // Products — StoreHub items only carry productId, name must be resolved.
-      for (const item of txn.items || []) {
-        const resolved = item.productId ? productNameById.get(item.productId) : null;
-        const name = resolved || item.name || null;
-        if (!name) continue;
-        const key = name.toLowerCase().trim();
-        if (!key) continue;
-        if (!productStats[key]) {
-          productStats[key] = { name, revenue: 0, quantity: 0, orders: 0 };
-        }
-        productStats[key].revenue += item.total;
-        productStats[key].quantity += item.quantity;
-        productStats[key].orders += 1;
-
-        // Per-round product stats
-        if (round) {
-          const rMap = productStatsByRound[round];
-          if (!rMap[key]) rMap[key] = { name, revenue: 0, quantity: 0, orders: 0 };
-          rMap[key].revenue += item.total;
-          rMap[key].quantity += item.quantity;
-          rMap[key].orders += 1;
-        }
-      }
-    }
-
-    // Build summary data for AI
-    // Collect all dates seen for blended target
-    const allDates7 = [...new Set(Object.values(roundStats7).flatMap(s => [...s.days]))].sort();
-    const allDates30 = [...new Set(Object.values(roundStats30).flatMap(s => [...s.days]))].sort();
-
-    // Targets are per-outlet, but revenue here is summed across all outlets
-    // pulled into this analysis. Scale the target by outlet count so the
-    // comparison is apples-to-apples.
-    const outletCount = Math.max(outlets.length, 1);
-
-    const roundSummary = ROUNDS.map((r) => {
-      const s7 = roundStats7[r.key];
-      const s30 = roundStats30[r.key];
-      const days7 = Math.max(s7.days.size, 1);
-      const days30 = Math.max(s30.days.size, 1);
-      const avgTarget7 = getAvgTarget(r.key, allDates7, activeTargets) * outletCount;
-      const avgTarget30 = getAvgTarget(r.key, allDates30, activeTargets) * outletCount;
-      const tEntry = activeTargets[r.key as keyof typeof activeTargets];
-      return {
-        round: r.label,
-        timeRange: `${r.startH}:00-${r.endH}:00`,
-        targetWeekday: tEntry.weekday.revenue,
-        targetWeekend: tEntry.weekend.revenue,
-        outletsIncluded: outletCount,
-        last7days: {
-          avgDailyRevenue: Math.round(s7.revenue / days7),
-          avgDailyOrders: Math.round((s7.orders / days7) * 10) / 10,
-          totalRevenue: Math.round(s7.revenue),
-          pctOfTarget: avgTarget7 > 0 ? Math.round((s7.revenue / days7 / avgTarget7) * 100) : 0,
-        },
-        last30days: {
-          avgDailyRevenue: Math.round(s30.revenue / days30),
-          avgDailyOrders: Math.round((s30.orders / days30) * 10) / 10,
-          totalRevenue: Math.round(s30.revenue),
-          pctOfTarget: avgTarget30 > 0 ? Math.round((s30.revenue / days30 / avgTarget30) * 100) : 0,
-        },
-      };
-    });
-
-    const totalRevenue30 = allTxns.reduce((s, t) => s + t.total, 0);
-    const totalOrders30 = allTxns.length;
-    const aov30 = totalOrders30 > 0 ? Math.round((totalRevenue30 / totalOrders30) * 100) / 100 : 0;
-
-    // Daily trend (last 7 days)
-    const last7Dates = Object.keys(dailyTotals).sort().slice(-7);
-    const dailyTrend = last7Dates.map((d) => ({
-      date: d,
-      revenue: Math.round(dailyTotals[d].revenue),
-      orders: dailyTotals[d].orders,
-    }));
-
-    // ─── Call Claude for analysis ───────────────────────────────────
-
-    const analysisPrompt = `You are a sales analyst for Celsius Coffee, a Malaysian specialty coffee chain. Analyze the following sales data and provide actionable recommendations.
-
-DATA CONTEXT:
-- Currency: Malaysian Ringgit (RM)
-- Business: Specialty coffee chain
-- Outlets: ${outlets.map((o) => o.name).join(", ")}
-- Analysis period: Last 30 days (${from30} to ${todayMYT})
-
-OVERALL METRICS (Last 30 days):
-- Total Revenue: RM ${Math.round(totalRevenue30).toLocaleString()}
-- Total Orders: ${totalOrders30.toLocaleString()}
-- Average Order Value (AOV): RM ${aov30}
-
-CHANNEL BREAKDOWN (Last 30 days):
-- Dine-In: RM ${Math.round(channelStats.dine_in.revenue)} (${channelStats.dine_in.orders} orders)
-- Takeaway: RM ${Math.round(channelStats.takeaway.revenue)} (${channelStats.takeaway.orders} orders)
-- Delivery: RM ${Math.round(channelStats.delivery.revenue)} (${channelStats.delivery.orders} orders)
-
-SALES BY ROUND (Daily target vs actual, AGGREGATED across ${outletCount} outlet${outletCount !== 1 ? "s" : ""}):
-Note: pctOfTarget below has ALREADY been scaled for outlet count — use these percentages directly.
-${roundSummary.map((r) => `${r.round} (${r.timeRange}): Per-outlet target Wkd RM${r.targetWeekday} / Wknd RM${r.targetWeekend} | Last 7d avg across all outlets: RM${r.last7days.avgDailyRevenue}/day (${r.last7days.pctOfTarget}% of combined target) | Last 30d avg: RM${r.last30days.avgDailyRevenue}/day (${r.last30days.pctOfTarget}%)`).join("\n")}
-
-DAILY TREND (Last 7 days):
-${dailyTrend.map((d) => `${d.date}: RM${d.revenue} (${d.orders} orders)`).join("\n")}
-
-TOP PRODUCTS PER ROUND (use these real product names in your recommendations):
-${ROUNDS.map((r) => {
-  const top = Object.values(productStatsByRound[r.key])
-    .sort((a, b) => b.revenue - a.revenue)
-    .slice(0, 5);
-  if (top.length === 0) return `${r.label}: (no product data)`;
-  return `${r.label}: ${top.map((p) => `${p.name} (${p.quantity}u, RM${Math.round(p.revenue)})`).join(" | ")}`;
-}).join("\n")}
-
-Return a JSON array of 3-5 SHARP, SPECIFIC recommendations. Each object:
-- "type": "opportunity" | "warning" | "insight" | "action"
-- "title": concrete and number-driven, max 60 chars
-- "description": 1-2 sentences. MUST cite specific numbers AND specific product names from the data above when relevant. Prescribe ONE concrete action.
-- "impact": "high" | "medium" | "low"
-- "category": "round_performance" | "product_mix" | "aov" | "channel" | "trend"
-
-Specificity rules (this is what makes insights useful, not generic):
-- Bad (too broad): "Implement breakfast combos to boost morning sales"
-- Good (specific): "Breakfast top seller is Flat White (45u/day, RM720). Bundle with Butter Croissant (currently 18u/day at breakfast) as 'Morning Set RM18' — lifts AOV from RM16 to RM18"
-- Bad: "Promote dinner items"
-- Good: "Dinner at 67% of target. Your top dinner item Iced Latte sells 32u at dinner vs 60u at lunch — run a 'Latte Hour' 7-9PM promo on it"
-- Every recommendation must name a specific product OR cite a specific number. No abstractions.
-- Prefer fewer sharp recs over many broad ones.
-- If a round is above target, skip it unless there's a specific scaling insight.
-
-Return ONLY valid JSON array, no markdown or explanation.`;
-
-    const response = await client.messages.create({
-      model: "claude-sonnet-4-6",
-      max_tokens: 1500,
-      messages: [{ role: "user", content: analysisPrompt }],
-    });
-
-    let recommendations: unknown[] = [];
-    try {
-      const text = response.content[0].type === "text" ? response.content[0].text : "";
-      // Parse JSON from response (handle potential markdown wrapping)
-      const jsonStr = text.replace(/```json\n?/g, "").replace(/```\n?/g, "").trim();
-      recommendations = JSON.parse(jsonStr);
-    } catch {
-      console.error("[sales/recommendations] Failed to parse AI response");
-      recommendations = [];
-    }
-
-    return NextResponse.json({
-      recommendations,
-      summary: {
-        totalRevenue30: Math.round(totalRevenue30),
-        totalOrders30,
-        aov30,
-        analysisDate: todayMYT,
-      },
-      targetsMeta,
-    });
-  } catch (err) {
-    console.error("[sales/recommendations] Error:", err);
-    return NextResponse.json({ error: "Internal server error" }, { status: 500 });
-  }
-}
